requirements.txt
pygame
pymunk
neat-python
typer
pydantic
pytest
tomli
